+++ tailwind.config.js
@@ -1,3 +1,5 @@
+const { fontFamily } = require('tailwindcss/defaultTheme');
+
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
@@ -5,7 +7,14 @@ module.exports = {
     './components/**/*.{js,ts,jsx,tsx}',
   ],
   theme: {
-    extend: {},
+    extend: {
+      fontFamily: {
+        "catamaran-light": ["Catamaran-Light"],
+        "catamaran-regular": ["Catamaran-Regular"],
+        "catamaran-medium": ["Catamaran-Medium"],
+        "catamaran-bold": ["Catamaran-Bold"],
+      },
+    },
   },
   plugins: [],
 }