+++ CHANGELOG.md
@@ -15,3 +15,4 @@ All notable changes to this project will be documented in this file.
 - user api endpoints
 - web-scalper
 - web-scalper api endpoints
+- Catamaran font family